requirements.txt
pynetworktables
opencv-python
cv2-enumerate-cameras
pygrabber
PyQt5
numpy
pywin32